refactor: change formatting
