+++ online-store/src/components/app/addResetButtonsFunctionality.ts
@@ -1,0 +1,11 @@
+/**
+ * Функционал кнопок сброса
+ */
+
+import { addResetFiltersButtonFunctionality } from './addResetFiltersButtonFunctionality';
+import { addResetStorageButtonFunctionality } from './addResetStorageButtonFunctionality';
+
+export const addResetButtonsFunctionality: () => void = () => {
+    addResetFiltersButtonFunctionality();
+    addResetStorageButtonFunctionality();
+};
