+++ online-store/src/components/app/setFocusOnSearchInput.ts
@@ -1,0 +1,8 @@
+/**
+ * Устанавливает фокус на поле поиска
+ */
+
+export const setFocusOnSearchInput: () => void = () => {
+    const searchInput = document.querySelector('.search-input') as HTMLInputElement;
+    searchInput.focus();
+};
